Make it a case class sos that equals works.
